change write_tga to take ints

+++ src/tga.h
@@ -17,6 +17,4 @@ int pix(int x, int y, int width, int height, int color);
  * the index of a pixel within the 1D char array is
  * given by the function pix above.
  */
-void write_tga(char *filename, char *pixels,
-               unsigned short width, unsigned short height,
-               int color);
+void write_tga(char *filename, char *pixels, int width, int height, int color);
